Moved getting SS key to discovery service. #135
