+++ agent/discovery/discovery.go
@@ -6,7 +6,9 @@ import (
 	"io/ioutil"
 	"net/http"
 	"os"
+	"os/exec"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/fromkeith/gossdp"
@@ -14,6 +16,7 @@ import (
 	"github.com/subutai-io/agent/config"
 	"github.com/subutai-io/agent/db"
 	"github.com/subutai-io/agent/lib/container"
+	"github.com/subutai-io/agent/lib/gpg"
 	"github.com/subutai-io/agent/lib/net"
 	"github.com/subutai-io/agent/log"
 )
@@ -115,4 +118,51 @@ func save(ip string) {
 		monitor.InitInfluxdb()
 	}
 	config.Management.Host = ip
+
+	if pk := getKey(); pk != nil {
+		gpg.ImportPk(pk)
+		config.Management.GpgUser = extractKeyID(pk)
+	}
+}
+
+func getKey() []byte {
+	client := &http.Client{Timeout: time.Second * 5}
+	if config.Management.Allowinsecure {
+		client = &http.Client{Transport: &http.Transport{TLSClientConfig: &tls.Config{InsecureSkipVerify: true}}, Timeout: time.Second * 5}
+	}
+	resp, err := client.Get("https://" + config.Management.Host + ":" + config.Management.Port + config.Management.RestPublicKey)
+	if log.Check(log.WarnLevel, "Getting Management host Public Key", err) {
+		return nil
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode == 200 {
+		if key, err := ioutil.ReadAll(resp.Body); err == nil {
+			return key
+		}
+	}
+
+	log.Warn("Failed to fetch PK from Management Server. Status Code " + strconv.Itoa(resp.StatusCode))
+	return nil
+}
+
+func extractKeyID(k []byte) string {
+	command := exec.Command("gpg")
+	stdin, err := command.StdinPipe()
+	if err != nil {
+		return ""
+	}
+
+	_, err = stdin.Write(k)
+	log.Check(log.DebugLevel, "Writing to stdin pipe", err)
+	log.Check(log.DebugLevel, "Closing stdin pipe", stdin.Close())
+	out, err := command.Output()
+	log.Check(log.WarnLevel, "Extracting ID from Key", err)
+
+	if line := strings.Fields(string(out)); len(line) > 1 {
+		if key := strings.Split(line[1], "/"); len(key) > 1 {
+			return key[1]
+		}
+	}
+	return ""
 }
